ReconfigurableLogger
